Added flag to pdf2txt to skip LTLayoutContainer.group_textboxes to speed up text generation on complex documents.
When working on documents there were several papers that took longer
than 5 mintues to convert to text. In order to speed up text generation,
the work performed to generate a hierarcy by the LTLayoutContainer.group_textboxes
can be skiped. This can be enabled by a command line flag.

Added use of this flag to the tests.

Added description of -S flag to html pages.

Signed-off-by: Brian D. Caruso <bdc34@cornell.edu>