Bugfix encodingdb: parsing hex instead of dec
A pdf file produced text elements with wrongly encoded text.

Based on this specification file:

    github.com/adobe-type-tools/agl-aglfn/blob/master/glyphlist.txt#L44

The unicode scalar values use hexadecimal values. This was confirmed by
the sample pdf that included names like `uni004F`. This did not result
in an error because the regex simply matched digital strings.

The encodingdb.py file was modified such that it now matches the last
four characters as hexadecimal characters (and otherwise fails) and parses
them as a hexadecimal number.

Tests on the pdf seemed to resolve the issue.

Signed-off-by: Willem Van Onsem <vanonsem.willem@gmail.com>